desktop/requirements-client.txt
PySide6==6.7.2
python-socketio==5.11.3
requests==2.32.3